// src/app/modules/admin/project/project-delivery-notes/edit-delivery-note-dialog/models/groupMPs.ts
import { InventoryItemDTO } from "app/modules/admin/inventory/models/inventoryItemDTO";

export interface GroupMPs {
  quoteId: number;
  cuttingListId: number;
  frameType: string;
  frameName: string;
  totalItems: number;
  totalLengthOrQty: number;
  groupedInventoryItemDTO: InventoryItemDTO[];
}